requirements.txt
# === Core Bot Frameworks ===
discord.py==2.4.0
aiogram==3.13.1

# === AI + API Clients ===
openai==1.51.0
httpx==0.27.2
requests==2.32.3

# === Async + Utility ===
aiohttp==3.10.10
aiohttp-socks==0.8.4
python-dotenv==1.0.1
apscheduler==3.10.4
pytz==2024.2

# === Data + Visualization ===
pandas==2.2.3
matplotlib==3.9.2
reportlab==4.2.5
